handle message call

+++ client/src/components/message/RightSide.js
@@ -133,8 +133,31 @@ const RightSide = () => {
    }, [isLoadMore])
 
    const handleDeleteConversation = () => {
-	   dispatch(deleteConversation({auth,id}))
-	   return history.push('/message')
+	   if(window.confirm('Do you want to delete?')) {
+		   dispatch(deleteConversation({auth,id}))
+		   return history.push('/message')
+	   }
+   }
+   
+   const caller = ({video}) => {
+	   const {_id, avatar, username, fullname} = user;
+	   const msg = {
+		   sender: auth.user._id,
+		   recipient: _id,
+		   avatar,
+		   username,
+		   fullname,
+		   video
+	   }
+	   dispatch({ type: GLOBAL_TYPES.CALL, payload: msg })
+	   
+   }
+   
+   const handleAudioCall = () => {
+	   caller({video: false})
+   }
+   const handleVideoCall = () => {
+	   caller({video: true})
    }
    
   return (
@@ -143,7 +166,12 @@ const RightSide = () => {
         {
           user.length !== 0 && (
             <UserCard user={user}>
-              <i className='fas fa-trash text-danger' onClick={handleDeleteConversation} />
+              <div>
+			  
+			  <i className='fas fa-phone-alt' onClick={handleAudioCall} />
+			  <i className='fas fa-video mx-3' onClick={handleVideoCall} />
+			  <i className='fas fa-trash text-danger' onClick={handleDeleteConversation} />
+			  </div>
             </UserCard>
           )
         }
